requirements.txt
asgiref==3.8.1
black==25.1.0
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
filelock==3.17.0
identify==2.6.8
inflection==0.5.1
isort==6.0.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
virtualenv==20.29.2
whitenoise==6.9.0
